Update Figure3.py
Clarified that Figure 9. also cannot be recreated.

From 2023 study in Section 3.5 "Occurrence of Extreme Precipitation Events":

"In the previous section, we focus on the analysis of the monthly intensity of extreme precipitation. Here, the monthly frequency of extreme precipitation is investigated with the same tree-based ML methods (RF and XGB) for the same regions."